squash! DefecTor countermeasure: change server- and client-side DNS TTL clipping
Clean up and repair based on conversations with pulls. I had thought
that one of these functions was about "what to report" and one was
about "what to believe", but they were about "what to report" and
"how long to cache at the exit" -- and they need to stay in sync.